add alignment + font weight
