+++ src/blocks/button-maxi/style.scss
@@ -1,6 +1,5 @@
 body.maxi-blocks--active .maxi-button-block {
 	display: flex !important;
-	flex-direction: column;
 	text-align: center;
 
 	&[data-align='full'] {
