+++ src/blocks/number-counter-maxi/style.scss
@@ -1,9 +1,9 @@
 body.maxi-blocks--active .maxi-number-counter-block {
 	// This element doesn't exist anymore
+	display: flex;
 	.maxi-number-counter {
 		&__box {
 			position: relative;
-			margin: 0 auto;
 			z-index: 2;
 			width: 100%;
 			max-width: 100%;
