+++ src/blocks/number-counter-maxi/styles.js
@@ -13,11 +13,15 @@ import {
 	getTransformStyles,
 	getZIndexStyles,
 	getFlexStyles,
+	getAlignmentFlexStyles,
 } from '../../extensions/styles/helpers';
 import { selectorsNumberCounter } from './custom-css';
 
 const getWrapperObject = props => {
 	const response = {
+		alignment: getAlignmentFlexStyles({
+			...getGroupAttributes(props, 'alignment'),
+		}),
 		opacity: getOpacityStyles({
 			...getGroupAttributes(props, 'opacity'),
 		}),
@@ -242,6 +246,7 @@ const getStyles = props => {
 			props
 		),
 	};
+	console.log(response);
 	return response;
 };
