+++ src/components/index.js
@@ -73,3 +73,4 @@ export { default as TransformControl } from './transform-control';
 export { default as TransitionControl } from './transition-control';
 export { default as TypographyControl } from './typography-control';
 export { default as ZIndexControl } from './zindex-control';
+export { default as FontWeightControl } from './font-weight-control';
