+++ src/components/typography-control/index.js
@@ -2,7 +2,7 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { select, dispatch, useSelect } from '@wordpress/data';
+import { dispatch, useSelect } from '@wordpress/data';
 import { useState } from '@wordpress/element';
 
 /**
@@ -16,8 +16,7 @@ import ResponsiveTabsControl from '../responsive-tabs-control';
 import SelectControl from '../select-control';
 import TextShadowControl from '../text-shadow-control';
 import SettingTabsControl from '../setting-tabs-control';
-import { loadFonts } from '../../extensions/text/fonts';
-
+import FontWeightControl from '../font-weight-control';
 import {
 	setFormat,
 	getCustomFormatValue,
@@ -35,7 +34,7 @@ import { getDefaultSCValue } from '../../extensions/style-cards';
  * External dependencies
  */
 import classnames from 'classnames';
-import { isNil, isBoolean, isNumber, isEmpty } from 'lodash';
+import { isBoolean, isNumber, isEmpty } from 'lodash';
 
 /**
  * Styles and icons
@@ -611,75 +610,7 @@ const TypographyControl = withFormatValue(props => {
 		return nonHoverValue;
 	};
 
-	const getWeightOptions = () => {
-		const { getFont } = select('maxiBlocks/text');
-
-		if (!isNil(getValue(`${prefix}font-family`))) {
-			const fontFiles = getFont(getValue(`${prefix}font-family`))?.files;
-			const fontOptions = fontFiles
-				? Object.keys(fontFiles).map(key => key)
-				: [];
-
-			if (fontOptions.length === 0) {
-				return [
-					{ label: __('Thin (Hairline)', 'maxi-blocks'), value: 100 },
-					{
-						label: __('Extra Light (Ultra Light)', 'maxi-blocks'),
-						value: 200,
-					},
-					{ label: __('Light', 'maxi-blocks'), value: 300 },
-					{
-						label: __('Normal (Regular)', 'maxi-blocks'),
-						value: 400,
-					},
-					{ label: __('Medium', 'maxi-blocks'), value: 500 },
-					{
-						label: __('Semi Bold (Semi Bold)', 'maxi-blocks'),
-						value: 600,
-					},
-					{ label: __('Bold', 'maxi-blocks'), value: 700 },
-					{
-						label: __('Extra Bold (Ultra Bold)', 'maxi-blocks'),
-						value: 800,
-					},
-					{ label: __('Black (Heavy)', 'maxi-blocks'), value: 900 },
-					{
-						label: __('Extra Black (Ultra Black)', 'maxi-blocks'),
-						value: 950,
-					},
-				];
-			}
-			const weightOptions = {
-				100: 'Thin (Hairline)',
-				200: 'Extra Light (Ultra Light)',
-				300: 'Light',
-				400: 'Normal (Regular)',
-				500: 'Medium',
-				600: 'Semi Bold (Semi Bold)',
-				700: 'Bold',
-				800: 'Extra Bold (Ultra Bold)',
-				900: 'Black (Heavy)',
-				950: 'Extra Black (Ultra Black)',
-			};
-			const response = [];
-			if (!fontOptions.includes('900')) {
-				fontOptions.push('900');
-			}
-			fontOptions.forEach(weight => {
-				const weightOption = {};
-				if (weightOptions[weight]) {
-					weightOption.label = __(
-						weightOptions[weight],
-						'maxi-blocks'
-					);
-					weightOption.value = weight;
-					response.push(weightOption);
-				}
-			});
-			return response;
-		}
-		return null;
-	};
+	//const getWeightOptions = () => {};
 	const getDefault = (prop, customBreakpoint) => {
 		const currentBreakpoint = customBreakpoint || breakpoint;
 
@@ -809,23 +740,16 @@ const TypographyControl = withFormatValue(props => {
 				!disableColor &&
 				!styleCards &&
 				!hideAlignment && <Divider />}
-			<SelectControl
-				label={__('Font weight', 'maxi-blocks')}
-				className='maxi-typography-control__weight'
-				value={getValue(`${prefix}font-weight`)}
-				options={getWeightOptions()}
+			<FontWeightControl
 				onChange={val => {
 					onChangeFormat({ [`${prefix}font-weight`]: val });
-					const fontName = getValue(`${prefix}font-family`);
-					const fontStyle = getValue(`${prefix}font-style`);
-					const objFont = { [fontName]: {} };
-
-					objFont[fontName].weight = val.toString();
-					if (fontStyle) objFont[fontName].style = fontStyle;
-
-					loadFonts(objFont);
 				}}
+				value={getValue(`${prefix}font-weight`)}
+				fontFamily={getValue(`${prefix}font-family`)}
+				fontStyle={getValue(`${prefix}font-style`)}
+				prefix={prefix}
 			/>
+
 			<SelectControl
 				label={__('Text transform', 'maxi-blocks')}
 				className='maxi-typography-control__transform'
